pyproject.toml
[tool.poetry]
name = "delivery_box"
version = "0.1.0"
description = ""
authors = ["Anton Sh. <trollknurr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.1.1"
hydra-colorlog = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.18.3"
pytest-mock = "^3.7.0"
black = "^22.3.0"
pylint = "^2.13.5"
isort = "^5.10.1"
mypy = "^0.942"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint.MASTER]
disable = ["C0114", "R0201", "R0902", "R0913",  "R0903"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
